render into viewport image

+++ src/renderer.h
@@ -18,7 +18,7 @@ namespace yart
         ///     The size of the array should be equal to width*height*3, where 3 denotes the number of channels in the output image (RGB)
         /// @param width Width in pixels of the output image
         /// @param height Height in pixels of the output image
-        void Render(double buffer[], uint32_t width, uint32_t height);
+        void Render(float buffer[], uint32_t width, uint32_t height);
 
     private:
     private:
